pyproject.toml
[tool.poetry]
name = "Todo-using-FastAPI"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.1"}
fastapi = {extras = ["all"], version = "^0.78.0"}
uvicorn = "^0.17.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
